fix (build): remove invalid import
Builds try to import debugging.TimestepInspector and attach it to the
connection, but the file doesn't exist, causing a jsio import error
during the build step. This looks like an error, but doesn't actually
break the build.

This change does not restore the TimestepInspector, it just removes the
attempted import.