fix (browser build): restructure splash image code
Change the reverse for loop that accesses every possible splash image
(was breaking when running into splash keys that did not exist in the
splash section of the manifest). Now the loop only covers potential
image keys that match the supportedOrientations and stops after finding
the first available one, with universal only accessed if nothing else is
found. This is not the complete logic necessary for handling splash
images: screen size and actual orientation should be taken into account.