improve createFileStream
onFile delayed the through2 callback till the implementor's promise
resolved, which blocked future writes in the stream. The correct
behavior is to call the callback immediately, only deferring the final
onFinish callback. The file stream can then accept all writes as soon as
they're available, and is therefore responsible for managing its own
concurrency.

onEnd renamed to onFinish to more accurately represent the fact that
it's called when the stream write finishes (corresponding to the
'finish' event on the write half of the stream).