add FileSink to the stream api as 'end-build'
Enables reuse of the build streams in custom builds.  For example, if
you wanted to sprite images outside of the normal build process, you
would create a build stream piping 'src' -> 'spriter' -> 'write-files'
-> 'end-build', where 'end-build' is the FileSink.

The 'end-build' stream has an `onFinishBuild` property that is a Promise
resolving to the build results object.