enable overriding of CONFIG items from CLI
To set a key in the JavaScript CONFIG object at build time, use a flag
starting with `--config:`:

  devkit release browser-mobile --config:<key> <value>

where <key> is a '.'-separated path and <value> is JSON parseable.

For example,

  devkit release browser-mobile --config:modules.analytics '{"id": 1}'

will set `CONFIG.modules.analytics = {id: 1}` (creating both the
`modules` and `analytics` objects if they do not already exist).