in release mode use a different bootstrap file that calls eval
The original bootstrap.js creates a script tag and sets the src for
loading the source.

This is nice for development because it plays well with the debuggers.
We want to do the opposite in production.  The bootstrap-production file
calls eval with no '//# sourceURL=' line so it does not show up in the
devtools.

A potential next step (prototyped but not included here) is to encrypt
the javascript and decrypt it locally.  This makes it less trivial to
grab the file off the wire.